修改文档

+++ run.sh
@@ -2,19 +2,26 @@
 
 WORK_DIR=/data/file/kongyu/dev
 
+# 输入参数
+SOURCE=$1
+
+if [ -z "${SOURCE}" ]
+then
+    echo "usge: $0 Ateng-Java"
+    exit 1
+fi
+
 set -x
 
 # 拉取代码
-git clone git@github.com:kongyu666/Ateng-Java.git ${WORK_DIR}/tmp/Ateng-Java
-git clone git@github.com:kongyu666/Ateng-Boot.git ${WORK_DIR}/tmp/Ateng-Boot
-git clone git@github.com:kongyu666/Ateng-Cloud.git ${WORK_DIR}/tmp/Ateng-Cloud
+git clone git@github.com:kongyu666/${SOURCE}.git ${WORK_DIR}/tmp/${SOURCE}
 
 # 删除旧代码
 rm -rf ${WORK_DIR}/work
 mv ${WORK_DIR}/tmp ${WORK_DIR}/work
 
 # 删除.git
-rm -rf ${WORK_DIR}/work/*/.git/
+rm -rf ${WORK_DIR}/work/${SOURCE}/.git/
 
 # 上传代码
 git add -v ${WORK_DIR}
